Dashboard part change to trait
