+++ app/Modules/CommissionModule/Presenters/CasesPresenter.php
@@ -2,6 +2,7 @@
 
 namespace PAF\Modules\CommissionModule\Presenters;
 
+use PAF\Common\BasePresenter;
 use PAF\Common\Feed\Components\Comment\CommentFeedControl;
 use PAF\Common\Feed\Components\FeedControl\FeedControlFactory;
 use PAF\Common\Lean\LeanSnapshots;
@@ -15,11 +16,13 @@ use Nette\Application\BadRequestException;
 use PAF\Modules\CommonModule\Components\CommentsControl\CommentsControl;
 use PAF\Modules\CommonModule\Components\CommentsControl\CommentsControlFactory;
 use PAF\Modules\CommonModule\Model\Comment;
-use PAF\Modules\CommonModule\Presenters\DashboardPresenter;
+use PAF\Modules\CommonModule\Presenters\Traits\DashboardComponent;
 use PAF\Modules\CommonModule\Services\CommentsService;
 
-final class CasesPresenter extends DashboardPresenter
+final class CasesPresenter extends BasePresenter
 {
+    use DashboardComponent;
+
     /** @var PafCases @inject */
     public $cases;
