+++ app/Modules/SettingsModule/Presenters/SettingsPresenter.php
@@ -2,11 +2,14 @@
 
 namespace PAF\Modules\SettingsModule\Presenters;
 
-use PAF\Modules\CommonModule\Presenters\DashboardPresenter;
+use PAF\Common\BasePresenter;
+use PAF\Modules\CommonModule\Presenters\Traits\DashboardComponent;
 use SeStep\GeneralSettings\Model\INode;
 
-final class SettingsPresenter extends DashboardPresenter
+final class SettingsPresenter extends BasePresenter
 {
+    use DashboardComponent;
+
     public function startup()
     {
         parent::startup();
